Refactor About You components into modular sections
Manual testing: Not run (container lacks browser UI).
